Added user model to for Authentication and Authorization

+++ routes/review.js
@@ -5,16 +5,28 @@ const router = express.Router({ mergeParams: true });
 const Listing = require("../models/listing.js");
 const Review = require("../models/review.js");
 const ExpressError = require("../utils/ExpressError.js");
+const User = require("../models/user.js");
+
+// ==== LOCAL MIDDLEWARES ====
+const { isLoggedIn } = require("../middleware.js");
 
 // ==== REVIEWS ROUTES ====
 
+//REVIEWS - REDIRECT ROUTE
+router.get("/", (req, res, next) => {
+  const { id } = req.params;
+  res.redirect(`/listings/${id}`);
+});
+
 //REVIEWS - CREATE ROUTE
-router.post("/", async (req, res, next) => {
+router.post("/", isLoggedIn, async (req, res, next) => {
   try {
     const { id } = req.params;
     const { review } = req.body;
+    const creatorId = req.user["_id"];
     let listing = await Listing.findById(id);
     review.createdAt = Date.now();
+    review.createdBy = await User.findById(creatorId);
     let newReview = new Review(review);
     await listing.reviews.push(newReview);
     await newReview.save();
@@ -28,16 +40,26 @@ router.post("/", async (req, res, next) => {
 
 //REVIEWS - DELETE ROUTE
 
-router.delete("/", async (req, res) => {
+router.delete("/", isLoggedIn, async (req, res, next) => {
   try {
     const { id } = req.params;
     const { listingId } = req.body;
-    await Listing.findByIdAndUpdate(listingId, {
-      $pull: { reviews: id },
-    });
-    await Review.findByIdAndDelete(id);
-    req.flash("success", "Review Deleted");
-    res.redirect(`/listings/${listingId}`);
+    const currUser = req.user["_id"];
+    let review = await Review.findById(id);
+    if (
+      review.createdBy["_id"].equals(currUser) ||
+      req.user.username == "admin"
+    ) {
+      await Listing.findByIdAndUpdate(listingId, {
+        $pull: { reviews: id },
+      });
+      await Review.findByIdAndDelete(id);
+      req.flash("success", "Review Deleted");
+      res.redirect(`/listings/${listingId}`);
+    } else {
+      req.flash("failure", "You cannot delete this review");
+      return res.redirect(`/listings/${listingId}`);
+    }
   } catch (err) {
     next(new ExpressError(400, "Review was not found"));
   }
